Error fixed
In some cases the algorithm would fail, it was not considered and the "floating point exception (core dumped)" appeared as a result. But now it all is in the past!
The code is still dirty and there are some pieces to work on, but the current version works as supposed.